WIP Slasher AI (FIX - VisAI Slasher is working now)
I moved around the Slasher AI's files then somehow the SlasherBP lost it's parent BP, which was the VisAI Controller BP. I made a new SlasherBP by repeating the same process as before. Got it working the same way I had it working the first time again.

The folder where I moved the previous files to "NEW-Slasher" is still in the content folder, but it is empty. I will make a new push commit where I remove it.