WIP Slasher V2
-Got Hit Reactions for Slasher to work again
-Added footsteps to Slasher
-Added a atmospheric test level where it's darker (try using your flashlight if you play)
-Reduced the current player melee weapon's damage values
-For now, Slasher AI can go into ragdoll death from all forms of player damage.

As we progress later on, we will keep it's ability to go into a ragdoll death for boss battles or game cutscenes only. However, as we further develop the Slasher's AI systems, it will detect if it is in a boss battle or cutscene, and if it is taking damage from melee weapons or more lethal weapons.