Grapple System WIP
Grapple interaction between player and Slasher AI is working, but it currently works manually.

The player can posses the Slasher AI or Amber's Player character and if they press 'G' then it will trigger the grapple. If nothing is pressed after couple of seconds then the grapple cancels, but that will not be in the final game since the slasher will always perform a grapple attack interaction if the player does not press an escape input on time.
Press 'G' again and the actor triggering the grapple will perform a grapple attack interaction. This presents how an escapable grapple will look like on player's escape failed from slasher.

Pressing 'Q' will perform an "escape" push interaction. This presents how an escapable grapple will look like on player's escape success from slasher. When the player escapes, then they are free from the grapple interaction, and the slasher is recovering from the push.