Grapple System WIP 2
Grapple interaction between player and Slasher AI is working, and it's triggered by the slasher AI.

-The slasher grapples when they are fully aggravated.
If the player is facing the slasher when the grapple is triggered then they will enter an escapable grapple, (currently) the player escapes by rotating. When they escape, the player pushes the slasher, and the slasher tries to recover, but at the moment they still move around during their recovery. If they player does not escape then they get stabbed twice.

-If the slasher is facing the player's back, then an unescapable grapple attack will occur.

-Added Frank Climax animations. Fighter and Stealth Takedown animations.