Extra Bug fix from speedboost
Same problem as before where if grapple failed because player moved too fast or it wasn't traced in time, but it's caused by the speedboost letting the player escape the trigger.

A better fix instead of just a band-aid fix is within the grapple event, force the AI to keep moving within a lesser acceptance radius. Even if it wasn't detected in time and the ai stayed still and broke, the player could still trigger a grapple if they walked into the trigger box.
Still not the solid fix, but a lesser band-aid fix.